Working towards getting the arrows to function properly
Getting the "clockwise" or "widdershins" part of the image in the MAIN card during match evaluation. Depending on whether the match is in a row or a column, appending the arrow direction to the imagename (as a number) and then returning the correct reading. 

Mostly there! Dupe data problems, and a glitch with the same imagename returning both up AND down for the same reading, which is not possible. :-/